Changes regarding public controller
Also renamed PublicFiles to Files
